# apps/web/src/lib/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RouteHub API",
    "version": "1.0.0",
    "description": "API documentation for RouteHub — fleet management, payments, documents, and inspections."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://geosafety.routehub.ge",
      "description": "Production (master)"
    },
    {
      "url": "https://team2.routehub.ge",
      "description": "Team 2"
    },
    {
      "url": "https://team3.routehub.ge",
      "description": "Team 3"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Supabase session cookie (set automatically after login)"
      },
      "cronSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Bearer token matching CRON_SECRET env var (for scheduled jobs)"
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/notifications/send-email": {
      "post": {
        "summary": "Send an email notification to a user",
        "description": "Looks up the user's email by user_id, generates an email from the notification type/title/message, and sends it via Resend. Accepts either a user session cookie or an x-internal-secret header (for server-side callers like cron jobs).\n",
        "tags": ["Notifications"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_id", "type", "title"],
                "properties": {
                  "user_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "type": {
                    "type": "string",
                    "description": "Notification type (e.g. item_overdue)"
                  },
                  "title": {
                    "type": "string"
                  },
                  "message": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object",
                    "description": "Additional context passed to email template"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sent": {
                      "type": "boolean"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields (user_id, type, or title)"
          },
          "401": {
            "description": "Authentication required (no session or internal secret)"
          },
          "404": {
            "description": "User email not found"
          },
          "500": {
            "description": "Failed to send email"
          }
        }
      }
    },
...
    "/api/inspectors": {
      "get": {
        "summary": "List inspectors",
        "description": "Returns all inspectors ordered by name. Optionally filter by status. Requires authentication.",
        "tags": ["Inspectors"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["active", "inactive", "all"]
            },
            "description": "Filter by status (omit or \"all\" to return every status)"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of inspector records"
          },
          "401": {
            "description": "Authentication required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create an inspector",
        "description": "Creates a new inspector record. Admin only.",
        "tags": ["Inspectors"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "is_active": {
                    "type": "boolean"
                  },
                  "vehicle_type": {
                    "type": "string"
                  },
                  "license_plate": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inspector created"
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "409": {
            "description": "Inspector with this email already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update an inspector",
        "description": "Updates an existing inspector by ID. Admin only.",
        "tags": ["Inspectors"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "full_name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "vehicle_type": {
                    "type": "string"
                  },
                  "license_plate": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inspector updated"
          },
          "400": {
            "description": "Validation failed or missing ID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete an inspector",
        "description": "Permanently deletes an inspector by ID. Admin only.",
        "tags": ["Inspectors"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Inspector ID to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Inspector deleted"
          },
          "400": {
            "description": "Missing inspector ID"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Admin access required"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}